+++ surfsense_backend/app/artifacts/media/podcast/__init__.py
@@ -1,0 +1,7 @@
+"""Podcast media → Artifact cutover."""
+
+from __future__ import annotations
+
+from app.artifacts.media.podcast.storage import open_stream, purge, persist
+
+__all__ = ["open_stream", "purge", "persist"]
